+++ src/services/logging/CoutLogger.cpp
@@ -10,7 +10,7 @@ void Ichor::CoutLogger::trace(const char *filename_in, int line_in, const char *
     if(_level <= LogLevel::LOG_TRACE) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
 
@@ -19,7 +19,7 @@ void Ichor::CoutLogger::debug(const char *filename_in, int line_in, const char *
     if(_level <= LogLevel::LOG_DEBUG) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
 
@@ -28,7 +28,7 @@ void Ichor::CoutLogger::info(const char *filename_in, int line_in, const char *f
     if(_level <= LogLevel::LOG_INFO) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
 
@@ -37,7 +37,7 @@ void Ichor::CoutLogger::warn(const char *filename_in, int line_in, const char *f
     if(_level <= LogLevel::LOG_WARN) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
 
@@ -46,7 +46,7 @@ void Ichor::CoutLogger::error(const char *filename_in, int line_in, const char *
     if(_level <= LogLevel::LOG_ERROR) {
         fmt::basic_memory_buffer<char, FMT_INLINE_BUFFER_SIZE> buf{};
         fmt::vformat_to(std::back_inserter(buf), format_str, args);
-        std::cout.write(buf.data(), buf.size()) << "\n";
+        std::cout.write(buf.data(), static_cast<int64_t>(buf.size())) << "\n";
     }
 }
